Split out controller tests into relevent epics modules
